requirements.txt
streamlit=1.24.0
flask=2.0.3
flask-cors=3.0.10
requests=2.26.0
dotenv=0.19.2
google-cloud-aiplatform=1.38.0